mv cli/genrator/tests -> tests/cli/generator
