+++ tests/fixtures/gen/deezer.json
