+++ tests/fixtures/gen/json_proto_mix_config.json
